fix duration

+++ app/src/main/java/com/task/webchallengetask/data/database/tables/ProgramTable.java
@@ -23,7 +23,7 @@ public class ProgramTable extends BaseModel {
     public String difficult;
 
     @Column
-    public int target;
+    public float target;
 
     @Column
     public String unit;
@@ -36,7 +36,7 @@ public class ProgramTable extends BaseModel {
         return difficult;
     }
 
-    public int getTarget() {
+    public float getTarget() {
         return target;
     }
